requirements.txt
Flask
requests
python-dotenv
pandas
openpyxl
gunicorn
